Turn and unit creation add and updates
It is now impossible to access to level up button and unit creation
button if the building or city isn't the owning of the current player.
Moreover, when the player click on the map, left and right menus are
reinitialized to the usual ones.
Moreover, turn management and unit creation management have been added.
However, there is still to add creation buttons for each unit.